Changed Computer, IOController interfaces to use Short/UShorts

+++ common/src/main/kotlin/common/Bytecode.kt
@@ -1,3 +1,3 @@
 package common
 
-typealias Bytecode = List<Int>
+typealias Bytecode = List<Short>
